modules/Module03/solution_pw_03/scripts/requirements.txt
# Зависимости для скрипта экспорта данных в CSV
# Установка: pip install -r requirements.txt

# Основные библиотеки для работы с данными
pandas>=1.5.0
numpy>=1.21.0

# Подключение к PostgreSQL
psycopg2-binary>=2.9.0

# Логирование и утилиты
python-dateutil>=2.8.0

# Для работы с путями и файлами
pathlib2>=2.3.0; python_version<"3.4"

# Дополнительные утилиты
openpyxl>=3.0.0  # Для работы с Excel файлами (опционально)
xlsxwriter>=3.0.0  # Для создания Excel отчетов (опционально)
